+++ src/components/cursor/constants.ts
