+++ src/components/map/points/commands/move-point.ts
@@ -1,9 +1,9 @@
-import type { Command } from "#app/history";
 import { ListNode, Store } from "#app/store";
 import type { StoreChangeEventKeys } from "#app/store/types";
 import type { LatLng } from "#app/types";
 import { UnifiedMap } from "#app/types/map";
 import { FireEvents } from "#components/map/helpers";
+import type { Command } from "#app/history/command";
 import { PointHelpers } from "../helpers";
 
 export class MovePointCommand implements Command {
